zpool clear: don't allow pool unsuspend if ZIL failed
`zpool clear` is a statement of intent from the operator: throw away
errors, get things going again. The problem is that if the if there are
callers blocked in fsync(), they will see the txg make it to disk, and
return success.

Fundamentally, this is the same old problem with fsync() and the ZIL: it
is assumed that txg sync completing is a 100% guarantee that all past
data is on disk.

The ZIL failure mechanism is not designed to deal with this case. The
ZIL still falls back to a txg sync if the ZIL write fails; the
difference is simply that it is keeping track of fsync() calls in
flight, so can (a) return EIO to them and (b) do itxg cleanup/callbacks
when the txg does eventually make it to disk.

However, ZIL failure is a good proxy to indicate that there were
fsync()s in progress when the pool suspended. So, this adds additional
checks to `zpool clear` to see if any datasets have a failed ZIL. If
they do, the request to unsuspend the pool is rejected. An escape hatch
switch is provided to override this behaviour.

The more correct fix is to have fsync()s in flight return EIO before as
the pool suspend, or after it returns. More thinking is required to
decide if that matches the expectation when failmode=wait is set.

An even better fix will be to use the information in memory (either the
cached DMU buffers, or the pending ZIL itxgs) to check or replay the
data, and return success to fsync() after the pool unsuspended if we can
be sure the data has made it to disk.

Both these options are significant work though. For now, this safety
switch will at least protect against a false sense of security.

Signed-off-by: Rob Norris <rob.norris@klarasystems.com>