test-runner: Fix signal propagation by creating a new process group when Popening a new test
The test-runner script offers a timeout mechanism, which kills
tests running for too long. This functionality has not worked properly
for some time. Here is a repro case:

1. Start killsnoop.bt to trace kill() signals on a Linux system
2. Configure a 5s timeout in a runfile.
3. Start a single test with ./scripts/zfs-tests.sh.

The output of killsnoop.bt shows SIGTERM is delivered but the process is
not killed. However, if we copy the PID of the process and issue a kill
command manually, the process will die. Here is the output of
killsnoop.bt:

        PID COMM              SIG    TPID RESULT
    3646956 kill               15 3644148      0

It turns out that sudo(8) is not going to propagate signals sent from
its own process group. That is why creating a new process group in
Popen() makes sudo actually propagate SIGTERM to its children.

Tracing the patched version of test-runner gives us the following
killsnoop.bt logs:

        PID COMM              SIG    TPID RESULT
    3626913 kill               15 3624372      0
    3624372 sudo               15 3624374      0
    3624372 sudo               15 3624372      0

Signed-off-by: Mateusz Piotrowski <mateusz.piotrowski@klarasystems.com>