Add a diagnostic kstat for obtaining pool status
This kstat output does not require taking the spa_namespace
lock, as in the case for 'zpool status'. It can be used for
investigations when pools are in a hung state while holding
global locks required for a traditional 'zpool status' to
proceed.

This kstat is not safe to use in conditions where pools are
in the process of configuration changes (i.e., adding/removing
devices). Therefore, this kstat is not intended to be a general
replacement or alternative to using 'zpool status'.

The newly introduced zfs_lockless_read_enabled module parameter
enables traversal of related kernel structures in cases where
the required read locks cannot be taken.

Sponsored-by: Klara, Inc.
Sponsored-by: Wasabi Technology, Inc.
Co-authored-by: Don Brady <don.brady@klarasystems.com>
Co-authored-by: Umer Saleem <usaleem@ixsystems.com>
Signed-off-by: Igor Ostapenko <igor.ostapenko@klarasystems.com>