zil: "crash" the ZIL if the pool suspends during fallback
Converts all txg_wait_synced() fallbacks to
txg_wait_synced_flags(TXG_WAIT_SUSPEND).

When this happens, it's necessary to:
- return all current waiters with an error
- disable the ZIL until the pool has returned

To do this, we add zil_crash(), which handles the bulk of this shutdown
work when something goes wrong. In particular, it immediately returns
error to all waiters in zil_commit(), and so to userspace.

The ZIL is marked for "restart" in 4 txgs, which naturally means some
time after the pool has resumed and enough txgs have passed to clean up
all the ZIL state. When this mark is set, new calls to zil_commit() will
be passed through to txg_wait_synced_flags(TXG_WAIT_SUSPEND), which will
either fail immediately if the pool is still suspended, or allow the
normal zil_commit() semantics to complete, just slowly.

Once the target txg restart is synced, zil_restart() is called to reopen
the ZIL.