zil: pass error back to itx callback
itx callbacks are used to signal that something can be cleaned up after
a itx is committed. Presently that's only used when syncing out mapped
pages (msync()) to mark dirty pages clean.

This makes it so that the error any error gets passed back to those
callbacks, so they can take a different cleanup action if necessary. For
mapped pages, that action is to re-mark them dirty, and also remove
their "eligible for reclaim" flag, as we do not want what may be the
only copy of a page to be evicted before the application can deal with
it.