zil: don't try to defer flushes if not attached to next zio
It's only safe to access the next LWB's vdev trace tree if we are
certain that the next LWB's write zio completion is not also accessing
it _and_ will not issue the flush before we are finished.

The only way to guarantee this is to ensure that the next LWB's zio
cannot enter the done function until this zio completes. We don't have a
reliable way to determine this from outside the zio pipeline, so we set
a flag when we attach the zios, and then check when considering whether
or not to defer.

Signed-off-by: Rob Norris <rob.norris@klarasystems.com>