Delegatable permission for `zfs redact`.
This changeset adds delegatable permission for `zfs redact`, as described
in https://github.com/openzfs/zfs/issues/10296

It adds "redact" as a permission, along with the code to enable its use,
and additions to the test suite.  It also adds mention of "redact" in
the man page for `zfs allow`.